move unmap method to FileSystem class
