+++ concourse-server/src/main/java/org/cinchapi/concourse/server/storage/temp/Buffer.java
@@ -55,7 +55,6 @@ import org.cinchapi.concourse.thrift.Operator;
 import org.cinchapi.concourse.thrift.TObject;
 import org.cinchapi.concourse.thrift.Type;
 import org.cinchapi.concourse.time.Time;
-import org.cinchapi.concourse.util.ByteBuffers;
 import org.cinchapi.concourse.util.Convert;
 import org.cinchapi.concourse.util.Logger;
 import org.cinchapi.concourse.util.NaturalSorter;
@@ -909,7 +908,7 @@ public final class Buffer extends Limbo {
          */
         public void delete() {
             FileSystem.deleteFile(filename);
-            ByteBuffers.unmap(content); // CON-163 (authorized)
+            FileSystem.unmap(content); // CON-163 (authorized)
             Logger.info("Deleting Buffer page {}", filename);
         }
