+++ concourse/src/main/java/org/cinchapi/concourse/util/ByteBuffers.java
@@ -23,9 +23,7 @@
  */
 package org.cinchapi.concourse.util;
 
-import java.lang.reflect.Method;
 import java.nio.ByteBuffer;
-import java.nio.MappedByteBuffer;
 import java.nio.charset.CharacterCodingException;
 import java.nio.charset.Charset;
 import java.nio.charset.CharsetDecoder;
@@ -258,25 +256,6 @@ public abstract class ByteBuffers {
             buffer.reset();
             return array;
         }
-    }
-
-    /**
-     * Attempt to force the unmapping of {@code buffer}. This method should be
-     * used with <strong>EXTREME CAUTION</strong>. If {@code buffer} is used
-     * after this method is invoked, it is likely that the JVM will crash.
-     * 
-     * @param buffer
-     */
-    // http://stackoverflow.com/a/19447758/1336833
-    public static void unmap(MappedByteBuffer buffer) {
-        try {
-            Method cleaner = buffer.getClass().getMethod("cleaner");
-            cleaner.setAccessible(true);
-            Method clean = Class.forName("sun.misc.Cleaner").getMethod("clean");
-            clean.setAccessible(true);
-            clean.invoke(cleaner.invoke(buffer));
-        }
-        catch (Exception e) {/* noop */}
     }
 
 }
